// backend/coding-challenge-back/src/DTO/album.dto.ts
export class AlbumDto {
    album: {
        album_type: string;
        artists: object[];
        available_markets: string[];
        external_urls: {
            spotify: string;
        }
        href: string;
        id: string;
        image: object[];
        name: string;
        release_date: string;
        release_date_precision: string;
        total_tracks: number;
        type: string;
        uri: string;
    };
    artists: {
        external_urls: object[],
        href: string;
        id: string;
        name: string;
        type: string;
        uri: string;
    }[];
    available_markets: string[];
    disc_number: number;
    duration_ms: number;
    explicit: boolean;
    external_ids: {
        isrc: string;
    };
    external_urls: {
        spotify: string;
    };
    href: string;
    id: string;
    is_local: boolean;
    name: string;
    popularity: number;
    preview_url: string;
    track_number: number;
    type: string;
    uri: string;
}