// backend/coding-challenge-back/src/DTO/artistAlbum.dto.ts
export class ArtistAlbumDto {
    album_group: string;
    album_type: string;
    artists: {
        external_urls: object[],
        href: string;
        id: string;
        name: string;
        type: string;
        uri: string;
    }[];
    available_markets: string[];
    external_urls: {
        spotify: string;
    };
    href: string;
    id: string;
    images: {
        height: number;
        url: string;
        width: number;
    }[];
    name: string;
    release_date: string;
    release_date_precision: string;
    total_tracks: number;
    type: string;
    uri: string;
}